Reorganize tax variables
