+++ openfisca_us/variables/gov/irs/tax/federal_income/capital_gains/dwks13.py
@@ -1,0 +1,22 @@
+from openfisca_us.model_api import *
+
+
+class dwks13(Variable):
+    value_type = float
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "DWKS13"
+    unit = USD
+    documentation = (
+        "search taxcalc/calcfunctions.py for how calculated and used"
+    )
+
+    def formula(tax_unit, period, parameters):
+        e24515 = add(tax_unit, period, ["unrecaptured_section_1250_gain"])
+        dwks11 = e24515 + add(
+            tax_unit, period, ["capital_gain_28_percent"]
+        )  # Sch D lines 18 and 19, respectively
+        dwks9 = tax_unit("dwks9", period)
+        dwks12 = min_(dwks9, dwks11)
+        dwks10 = tax_unit("dwks10", period)
+        return (dwks10 - dwks12) * tax_unit("hasqdivltcg", period)
