+++ openfisca_us/variables/gov/irs/tax/federal_income/capital_gains/dwks14.py
@@ -1,0 +1,17 @@
+from openfisca_us.model_api import *
+
+
+class dwks14(Variable):
+    value_type = float
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "DWKS14"
+    unit = USD
+    documentation = (
+        "search taxcalc/calcfunctions.py for how calculated and used"
+    )
+
+    def formula(tax_unit, period, parameters):
+        dwks1 = tax_unit("taxable_income", period)
+        dwks13 = tax_unit("dwks13", period)
+        return max_(0, dwks1 - dwks13) * tax_unit("hasqdivltcg", period)
