+++ openfisca_us/variables/gov/irs/tax/federal_income/income_tax_before_refundable_credits.py
@@ -20,22 +20,3 @@ class c09200(Variable):
 income_tax_before_refundable_credits = variable_alias(
     "income_tax_before_refundable_credits", c09200
 )
-
-
-class iitax(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    unit = USD
-    label = "Federal income tax"
-    documentation = "Total federal individual income tax liability."
-
-    def formula(tax_unit, period, parameters):
-        before_refundable_credits = tax_unit(
-            "income_tax_before_refundable_credits", period
-        )
-        refundable_credits = tax_unit("income_tax_refundable_credits", period)
-        return before_refundable_credits - refundable_credits
-
-
-income_tax = variable_alias("income_tax", iitax)
